chore: rename ops/boxes -> box, ops/points -> point (singular consistency)
