+++ packages/react-native-executorch/src/extensions/cv/ops/index.ts
@@ -2,6 +2,6 @@
  * Aggregated image, box, point, and quad operations.
  */
 export * from './image';
-export * from './boxes';
-export * from './points';
+export * from './box';
+export * from './point';
 export * from './quad';
